Conflicts:
	gromos++/src/utils/Hbond_calc.cc
	gromos++/src/utils/Hbond_calc.h

+++ gromos++/programs/epath.cc
@@ -443,7 +443,7 @@ int main(int argc, char **argv) {
             withH.addSpecifier("a:a");
 
             // remove H
-            for (int i = 0; i < withH.size(); i++) {
+            for (unsigned int i = 0; i < withH.size(); i++) {
                 if (!sys.mol(withH.mol(i)).topology().atom(withH.atom(i)).isH()) {
                     allatoms.addAtom(withH.mol(i), withH.atom(i));
                 }
@@ -547,7 +547,7 @@ int main(int argc, char **argv) {
                 }
 
                 bool domore = true;
-                while ((int)visited.size() != allatoms.size() && domore) {
+                while (visited.size() != allatoms.size() && domore) {
                     ///////////////////
 
                     // create specifier pairlist
@@ -567,7 +567,7 @@ int main(int argc, char **argv) {
 
                     Neighbours nb(sys, current.mol(0), current.atom(0));
                     // put the pairlist into the pair specifier
-                    for (int i = 0; i < around.size(); ++i) {
+                    for (unsigned int i = 0; i < around.size(); ++i) {
                         // check if atom is visited
                         if (!sys.mol(around.mol(i)).topology().atom(around.atom(i)).isH()
                                 && !visited.count(allatoms.findAtom(around.mol(i), around.atom(i)))) {
@@ -592,7 +592,7 @@ int main(int argc, char **argv) {
                     for (unsigned int i = 0; i < nb.size(); i++) {
                         if (sys.mol(current.mol(0)).topology().atom(nb[i]).isH()) {
                             // now search in the pairs list for a suitable acceptor
-                            for (int j = 0; j < pairs.size(); j++) {
+                            for (unsigned int j = 0; j < pairs.size(); j++) {
                                 // loop over masses
                                 bool couldbeh = false;
                                 for (unsigned int h = 0; h < masses.size(); h++) {
@@ -625,7 +625,7 @@ int main(int argc, char **argv) {
                         }
                     }
                     if (couldbeh) {
-                        for (int i = 0; i < pairs.size(); i++) {
+                        for (unsigned int i = 0; i < pairs.size(); i++) {
                             Neighbours pnb(sys, pairs.mol(i), pairs.atom(i));
                             for (unsigned int j = 0; j < pnb.size(); j++) {
                                 if (sys.mol(pairs.mol(i)).topology().atom(pnb[j]).isH()) {
@@ -646,7 +646,7 @@ int main(int argc, char **argv) {
                     }
 
                     // remove the hb from the pairs
-                    for (int i = 0; i < hb.size(); i++) {
+                    for (unsigned int i = 0; i < hb.size(); i++) {
                         pairs.removeAtom(hb.mol(i), hb.atom(i));
                     }
 
@@ -671,7 +671,7 @@ int main(int argc, char **argv) {
                     int jtype;
 
                     // covalent
-                    for (int i = 0; i < cov.size(); i++) {
+                    for (unsigned int i = 0; i < cov.size(); i++) {
                         // current in the donor or acceptor range?
                         if (((start.findAtom(cov.mol(i), cov.atom(i)) != -1) && (start.findAtom(current.mol(0), current.atom(0)) != -1))|| ((stop.findAtom(cov.mol(i), cov.atom(i)) != -1) && (stop.findAtom(current.mol(0), current.atom(0)) != -1))) {
                             ec = currentdecay * 1.00;
@@ -696,7 +696,7 @@ int main(int argc, char **argv) {
                     }
 
                     // hbond
-                    for (int i = 0; i < hb.size(); i++) {
+                    for (unsigned int i = 0; i < hb.size(); i++) {
                         // current in the donor or acceptor range?
                         if (((start.findAtom(hb.mol(i), hb.atom(i)) != -1) && (start.findAtom(current.mol(0), current.atom(0)) != -1))|| ((stop.findAtom(hb.mol(i), hb.atom(i)) != -1) && (stop.findAtom(current.mol(0), current.atom(0)) != -1))) {
                             eh = currentdecay * 1.00;
@@ -721,7 +721,7 @@ int main(int argc, char **argv) {
                     }
 
                     // space
-                    for (int i = 0; i < pairs.size(); i++) {
+                    for (unsigned int i = 0; i < pairs.size(); i++) {
                         // current in the donor or acceptor range?
                         if (((start.findAtom(pairs.mol(i), pairs.atom(i)) != -1) && (start.findAtom(current.mol(0), current.atom(0)) != -1))|| ((stop.findAtom(pairs.mol(i), pairs.atom(i)) != -1) && (stop.findAtom(current.mol(0), current.atom(0)) != -1))) {
                             es = currentdecay * 1.00;
@@ -951,7 +951,7 @@ int main(int argc, char **argv) {
                     int atomoffset = 0;
                     int m = 0;
                     int atomsoff = 0;
-                    for (int i = 0; i < allatoms.size(); i++) {
+                    for (unsigned int i = 0; i < allatoms.size(); i++) {
 
                         if (allatoms.mol(i) > m) {
                             resoffset += sys.mol(allatoms.mol(i) - 1).topology().numRes();
@@ -987,7 +987,7 @@ int main(int argc, char **argv) {
 
 
                     // Chain B
-                    for (int i = 0; i < path.size(); i++) {
+                    for (unsigned int i = 0; i < path.size(); i++) {
                         atomsoff = 0;
                         // complicated way of doing the resoffset
                         resoffset = 0;
@@ -1052,7 +1052,7 @@ int main(int argc, char **argv) {
                     int atomoffset_connect1plus = 0;
                     int atomoffset_plus1 = 0;
 
-                    for (int i = 0; i < path.size() - 1; i++) {
+                    for (unsigned int i = 0; i < path.size() - 1; i++) {
                         // complicated way of doing the offset
                         // reset
                         atomoffset_connect1 = 0;
